completed/04_지원서_평가_에이전트_(루브릭_기반_설명_가능_평가)/requirements.txt
anthropic>=0.39.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
pydantic>=2.0.0
matplotlib>=3.8.0
numpy>=1.26.0
tenacity>=8.2.0
chardet>=5.2.0
